requirements.txt
Flask
Pillow
cairosvg
requests
flask-cors
